EatFitService/requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.3.0.23
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==1.11.8
django-celery==3.1.17
django-cleanup==1.0.1
django-pyodbc-azure==1.11.0.0
djangorestframework==3.6.3
fuzzywuzzy==0.10.0
idna==2.6
kombu==4.1.0
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
pip==9.0.1
pyodbc==4.0.21
pytz==2016.4
requests==2.18.4
setuptools==15.0
suds==0.4
urllib3==1.22
vine==1.1.4
wheel==0.24.0
xlrd==0.9.4
xlutils==1.7.1
xlwt==1.0.0
